+++ app/src/main/java/com/fieldbook/tracker/brapi/service/BrAPIServiceV2.java
@@ -1545,7 +1545,7 @@ public class BrAPIServiceV2 extends AbstractBrAPIService implements BrAPIService
             field.setCount(studyDetails.getNumberOfPlots().toString());
             field.setObservation_level(observationLevel);
             field.setImport_format(ImportFormat.BRAPI);
-
+            field.setTrial_name(studyDetails.getTrialName());
             // Get our host url
             if (BrAPIService.getHostUrl(context) != null) {
                 field.setExp_source(BrAPIService.getHostUrl(context));