+++ app/src/main/java/com/fieldbook/tracker/traits/formats/MultiCategoricalFormat.kt
@@ -20,4 +20,4 @@ class MultiCategoricalFormat : TraitFormat(
     NameParameter(),
     DetailsParameter(),
     CategoriesParameter()
-), StringCoder by CategoricalJsonCoder(), ValuePresenter by CategoricalValuePresenter()
+), StringCoder by CategoricalJsonCoder(), ValuePresenter by CategoricalValuePresenter(), Scannable